04_concepts/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.0
h5py==3.9.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
nibabel==5.1.0
numpy==1.23.5
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
zennit==0.4.6
zennit-crp @ git+https://github.com/rachtibat/zennit-crp.git@c1f35ed551203993479a321e010bad67211ed0ab
